frames_straming/requirements.txt
opencv-python
librosa 
async-redis



